+++ cmd/dfuseeos/cli/dashboard.go
@@ -18,6 +18,7 @@ func init() {
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("dashboard-grpc-listen-addr", DashboardGrpcServingAddr, "TCP Listener addr for http")
 			cmd.Flags().String("dashboard-http-listen-addr", DashboardHTTPListenAddr, "TCP Listener addr for gRPC")
+			cmd.Flags().String("dashboard-metrics-api-addr", "http://127.0.0.1"+MetricsListenAddr, "HTTP address where to reach the metrics API endpoint")
 			cmd.Flags().String("dashboard-eos-node-manager-api-addr", EosManagerAPIAddr, "Address of the superviser manager api")
 			return nil
 		},
@@ -26,10 +27,11 @@ func init() {
 				HTTPListenAddr:      viper.GetString("dashboard-http-listen-addr"),
 				GRPCListenAddr:      viper.GetString("dashboard-grpc-listen-addr"),
 				NodeManagerAPIAddr:  viper.GetString("dashboard-eos-node-manager-api-addr"),
+				MetricsHTTPAddr:     viper.GetString("dashboard-metrics-api-addr"),
 				DmeshServiceVersion: viper.GetString("search-common-mesh-service-version"),
 				Title:               "dfuse for EOSIO - dashboard",
 				BlockExplorerName:   "eosq",
-				HeadBlockNumberApp:  "mindreader-node",
+				HeadBlockNumberApp:  "mindreader",
 			}, &dashboard.Modules{
 				Launcher:    modules.Launcher,
 				DmeshClient: modules.SearchDmeshClient,