+++ cmd/dfuseeos/cli/mindreader-stdin.go
@@ -9,7 +9,9 @@ import (
 	pbcodec "github.com/dfuse-io/dfuse-eosio/pb/dfuse/eosio/codec/v1"
 	"github.com/dfuse-io/dlauncher/launcher"
 	"github.com/dfuse-io/logging"
+	nodeManager "github.com/dfuse-io/node-manager"
 	nodeMindreaderStdinApp "github.com/dfuse-io/node-manager/app/node_mindreader_stdin"
+	"github.com/dfuse-io/node-manager/metrics"
 	"github.com/dfuse-io/node-manager/mindreader"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -19,13 +21,12 @@ import (
 
 func init() {
 	appLogger := zap.NewNop()
-
 	logging.Register("github.com/dfuse-io/dfuse-eosio/mindreader_stdin", &appLogger)
 
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "mindreader-stdin",
-		Title:       "deep-mind reader from stdin",
-		Description: "deep-mind reader from stdin, does not start nodeos itself",
+		Title:       "deep-mind reader node (stdin)",
+		Description: "Blocks reading node, unmanaged, reads deep mind from standard input",
 		MetricsID:   "mindreader-stdin",
 		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/mindreader_stdin$", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
 		RegisterFlags: func(cmd *cobra.Command) error {
@@ -52,16 +53,26 @@ func init() {
 				return codec.BlockFromProto(blk)
 			}
 
+			metricID := "mindreader-stdin"
+			headBlockTimeDrift := metrics.NewHeadBlockTimeDrift(metricID)
+			headBlockNumber := metrics.NewHeadBlockNumber(metricID)
+			metricsAndReadinessManager := nodeManager.NewMetricsAndReadinessManager(headBlockTimeDrift, headBlockNumber, viper.GetDuration("mindreader-readiness-max-latency"))
+
 			return nodeMindreaderStdinApp.New(&nodeMindreaderStdinApp.Config{
+				GRPCAddr:                   viper.GetString("mindreader-grpc-listen-addr"),
 				ArchiveStoreURL:            archiveStoreURL,
 				MergeArchiveStoreURL:       mergeArchiveStoreURL,
 				MergeUploadDirectly:        viper.GetBool("mindreader-merge-and-store-directly"),
-				GRPCAddr:                   viper.GetString("mindreader-grpc-listen-addr"),
 				MindReadBlocksChanCapacity: viper.GetInt("mindreader-blocks-chan-capacity"),
+				StartBlockNum:              viper.GetUint64("mindreader-start-block-num"),
+				StopBlockNum:               viper.GetUint64("mindreader-stop-block-num"),
+				DiscardAfterStopBlock:      viper.GetBool("mindreader-discard-after-stop-num"),
+				FailOnNonContinuousBlocks:  viper.GetBool("mindreader-fail-on-non-contiguous-block"),
 				WorkingDir:                 mustReplaceDataDir(dfuseDataDir, viper.GetString("mindreader-working-dir")),
 			}, &nodeMindreaderStdinApp.Modules{
-				ConsoleReaderFactory:     consoleReaderFactory,
-				ConsoleReaderTransformer: consoleReaderBlockTransformer,
+				ConsoleReaderFactory:       consoleReaderFactory,
+				ConsoleReaderTransformer:   consoleReaderBlockTransformer,
+				MetricsAndReadinessManager: metricsAndReadinessManager,
 			}, appLogger), nil
 		},
 	})