+++ CHANGELOG.md
@@ -11,6 +11,8 @@ date.
 
 ### Added
 
+* Added `--metrics-listen-addr` to control on which address to server the metrics API (Prometheus), setting this value to an empty string disable metrics serving.
+* Added `--dashboard-metrics-api-addr` to specify a different API address where to retrieve metrics for the dashboard.
 * Experimental support for `netkv://127.0.0.1:1234` as a possible `kvdb` database backend, which allows decoupling of single pods deployment into using an extremely simple networked k/v store, using the same badger backend and database as when you boot with default parameters.